zScore test tables

+++ dhis_android_analytics/src/main/java/dhis2/org/analytics/charts/data/Graph.kt
@@ -66,7 +66,7 @@ data class SerieData(
 
 data class GraphPoint(
     val eventDate: Date,
-    val position: Int? = -1,
+    val position: Float? = -1f,
     val fieldValue: Float
 )
